Composite multiple independent transform property animations
Previously, individual transform properties were lumped together with
transform lists, and only a single transform could be composited. With
this patch, we support multiple transforms as long as each affects a
different property (transform, translate, rotate, or scale).

Reland of
https://chromium-review.googlesource.com/c/chromium/src/+/3016295

With fixes to address performance regressions that were introduced in
the original patch.

1. Restored the original set of target properties. The transform
   type is now carried as part of the target property ID in the keyframe
   model.

2. Introduce a fast path. Transforms can be directly applied if only
   one transform type affects an element.

3. Avoid having an extra map for element animations requiring a
   compound transformation.

With these optimization, we still see a performance hit on the
targeted test; however, the impact has been greatly reduced.

Baseline:
https://pinpoint-dot-chromeperf.appspot.com/job/1440aa5a320000

Post optimization:
https://pinpoint-dot-chromeperf.appspot.com/job/11e374fe320000

Bug: 696374
Change-Id: I5cc20d0800f71f41b415b5fc57ed4a7fa10c525a