Hard-code the list of plugins and mimetypes in navigator
See [1] for a previous attempt to completely empty the navigator.plugins
and navigator.mimeTypes APIs. That caused site breakage due to sites
scanning for a PDF reader. This new attempt is discussed in [2], and
involves the hard-coding of a list of PDF viewers and mime types. The
list will be empty if the user setting to download PDFs instead of
viewing them (chrome://settings/content/pdfDocuments) is enabled.

Tests were copied mostly verbatim from [3], thanks @domenic.

I2S:
https://groups.google.com/a/chromium.org/g/blink-dev/c/bbxAGu90LgM

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2783393
[2] https://github.com/whatwg/html/pull/6738
[3] https://github.com/web-platform-tests/wpt/pull/29559

Bug: 1164635
Change-Id: I7c52af5b918768d8b4c4a9faa409fb4e6b72ecc2