[LayoutNG] Avoid using legacy positioned objects set
This patch improves performance of perf_tests/layout/abspos.html
by around 4%.
The key part to get that performance gain is to avoid calling
InsertIntoLegacyPositionedObjects()
in NGOutOfFlowLayoutPart::LayoutCandidates().

In order to do this, we need to implement extra changes
in other parts of the code.
The patch adds a new flag has_out_of_flow_descendants
in NGPhysicalContainerFragment, which will indicate
if there are positioned descendants that we need to process
(though we don't have a set to iterate them anymore we need
to traverse the tree).
The main changes are in LayoutBlock::RemovePositionedObjects()
which now needs to traverse down the tree looking
for positioned descendants.
There are also changes in ScrollAnchor::FindAnchorRecursive()
so we traverse the fragment tree instead of using the legacy
positioned objects list for LayoutNG objects.

Bug: 1098231
Change-Id: I50403616ccd3f0e34e9589445bb5c4cd14cdf6c0