AppHistoryNavigateEvent.navigationType, AppHistoryDestination.key, and AppHistoryDestination.id
navigationType is set to one of ("push", "replace", "reload", "traverse"), and
is derived from the WebFrameLoadType of the navigation.

key and id are set on AppHistoryDestination only in the "traverse" case,
because that is the only case where the key and id are known at the time
of the navigate event.

Bug: 1183545
Change-Id: I772418351931c3a74a45410001b51ea737da61a8